+++ src/main/java/application/FormController.java
@@ -36,7 +36,7 @@ import java.util.HashMap;
 @Api(value = "sessioncontroller", description = "Operations for navigating CommCare Menus, Cases, and Forms")
 @RestController
 @EnableAutoConfiguration
-public class SessionController {
+public class FormController {
 
     @Autowired
     private SessionRepo sessionRepo;
@@ -47,13 +47,7 @@ public class SessionController {
     @Autowired
     private RestoreService restoreService;
 
-    @Autowired
-    private MenuRepo menuRepo;
-
-    @Autowired
-    private InstallService installService;
-
-    Log log = LogFactory.getLog(SessionController.class);
+    Log log = LogFactory.getLog(FormController.class);
     ObjectMapper mapper = new ObjectMapper();
 
     @ApiOperation(value = "Start a new form entry session")
@@ -198,75 +192,11 @@ public class SessionController {
         return new SyncDbResponseBean();
     }
 
-    @ApiOperation(value = "Install the application at the given reference")
-    @RequestMapping(value = Constants.URL_INSTALL, method = RequestMethod.POST)
-    public SessionBean performInstall(@RequestBody InstallRequestBean installRequestBean) throws Exception {
-        InstallRequest installRequest = new InstallRequest(installRequestBean, restoreService, menuRepo, installService);
-        return getNextMenu(installRequest.getMenuSession(), true);
-    }
-
-    /**
-     * Make a menu selection, return a new Menu or a Form to display. This form will load a MenuSession object
-     * from persistence, make the selection, and update the record. If the selection began form entry, this will
-     * also create a new FormSession object.
-     *
-     * @param menuSelectBean Give the selection to be made on the current MenuSession
-     *                       (could be a module, form, or case selection)
-     * @return A MenuResponseBean or a NewFormSessionResponse
-     * @throws Exception
-     */
-    @ApiOperation(value = "Make the given menu selection and return the next set of options, or a form to play.")
-    @RequestMapping(value = Constants.URL_MENU_SELECT, method = RequestMethod.POST)
-    public SessionBean selectMenu(@RequestBody MenuSelectBean menuSelectBean) throws Exception {
-        MenuSession menuSession = getMenuSession(menuSelectBean.getSessionId());
-        boolean redrawing = menuSession.handleInput(menuSelectBean.getSelection());
-        menuRepo.save(menuSession.serialize());
-        return getNextMenu(menuSession, redrawing);
-    }
-
-    private SessionBean getNextMenu(MenuSession menuSession, boolean redrawing) throws Exception {
-
-        OptionsScreen nextScreen;
-
-        // If we were redrawing, remain on the current screen. Otherwise, advance to the next.
-        if(redrawing){
-            nextScreen = menuSession.getCurrentScreen();
-        }else{
-            nextScreen = menuSession.getNextScreen();
-        }
-
-        // No next menu screen? Start form entry!
-        if (nextScreen == null){
-            return menuSession.startFormEntry(sessionRepo);
-        }
-        else{
-            MenuResponseBean menuResponseBean = new MenuResponseBean();
-            menuResponseBean.setSessionId(menuSession.getSessionId());
-            // We're looking at a module or form menu
-            if(nextScreen instanceof MenuScreen){
-                MenuScreen menuScreen = (MenuScreen) nextScreen;
-                menuResponseBean.setMenuType(Constants.MENU_MODULE);
-                menuResponseBean.setOptions(getMenuRows(menuScreen));
-            }
-            // We're looking at a case list or detail screen (probably)
-            else if (nextScreen instanceof EntityScreen) {
-                EntityScreen entityScreen = (EntityScreen) nextScreen;
-                menuResponseBean.setMenuType(Constants.MENU_ENTITY);
-                menuResponseBean.setOptions(getMenuRows(entityScreen.getCurrentScreen()));
-            }
-            return menuResponseBean;
-        }
-    }
-
     private void updateSession(FormSession formEntrySession, SerializableFormSession serialSession) throws IOException {
         serialSession.setFormXml(formEntrySession.getFormXml());
         serialSession.setInstanceXml(formEntrySession.getInstanceXml());
         serialSession.setSequenceId(formEntrySession.getSequenceId() + 1);
         sessionRepo.save(serialSession);
-    }
-
-    private MenuSession getMenuSession(String sessionId) throws Exception {
-        return new MenuSession(menuRepo.find(sessionId), restoreService, installService);
     }
 
     private HashMap<Integer, String> getMenuRows(OptionsScreen nextScreen){
